+++ src/types.ts
@@ -1,3 +1,6 @@
+import { CustomStateSet } from "./CustomStateSet";
+import { ElementInternals } from "./element-internals";
+
 export interface IAom {
   ariaAtomic: string;
   ariaAutoComplete: string;
@@ -40,24 +43,36 @@ export interface IAom {
 export interface IElementInternals extends IAom {
   checkValidity: () => boolean;
   form: HTMLFormElement;
-  labels: NodeListOf<HTMLLabelElement>|[];
+  labels: NodeListOf<HTMLLabelElement> | [];
   reportValidity: () => boolean;
-  setFormValue: (value: string | FormData) => void;
-  setValidity: (validityChanges: Partial<globalThis.ValidityState>, validationMessage?: string, anchor?: HTMLElement) => void;
+  setFormValue: (value: string | FormData | null) => void;
+  setValidity: (
+    validityChanges: Partial<globalThis.ValidityState>,
+    validationMessage?: string,
+    anchor?: HTMLElement
+  ) => void;
+  shadowRoot: ShadowRoot|null;
+  states: CustomStateSet;
   validationMessage: string;
   validity: globalThis.ValidityState;
   willValidate: boolean;
 }
 
 export interface ICustomElement extends HTMLElement {
-  attributeChangedCallback?: (name: string, oldValue: any, newValue: any) => void;
-  connectedCallback?: () => void;
-  disconnectedCallback?: () => void;
-  attachedCallback?: () => void;
-  formDisabledCallback?: (isDisabled: boolean) => void;
-  formResetCallback?: () => void;
-  formAssociatedCallback?: (form: HTMLFormElement) => void;
+  constructor: (...args: any[]) => HTMLElement;
+  attributeChangedCallback(
+    name: string,
+    oldValue: any,
+    newValue: any
+  ): void;
+  connectedCallback(): void;
+  disconnectedCallback(): void;
+  attachedCallback(): void;
+  attachInternals(): ElementInternals;
+  formDisabledCallback(isDisabled: boolean): void;
+  formResetCallback(): void;
+  formAssociatedCallback(form: HTMLFormElement): void;
   disabled?: boolean;
 }
 
-export type LabelsList = NodeListOf<HTMLLabelElement>|[];
+export type LabelsList = NodeListOf<HTMLLabelElement> | [];